fix: resolve 17 test failures across preprocessing, API, and feature engineering
- ingestion.py: add card1 column to synthetic data generator (1000-1050 range
  with repeats) — velocity and amount aggregation features silently skipped
  without it (if "card1" not in df.columns: return df)
- preprocessing.py: drop remaining object columns (M1, M2, M3, DeviceType)
  before SMOTE — SMOTE requires all-numeric input, string columns cause
  ValueError; track _obj_cols_to_drop and apply same drop in transform()
- preprocessing.py: at inference (fit=False), fill missing numeric cols with
  NaN and missing cat cols with "missing" before calling imputers — sklearn
  imputers require exactly the fitted feature set; partial inference inputs
  were crashing with "Feature names seen at fit time, yet now missing"
- preprocessing.py: store _fit_col_order during fit_transform and reindex
  at transform time — missing cols added to df_eng at inference were appended
  at the end, causing "Feature names must be in same order" error in sklearn
- build_features.py: fix FutureWarning in _target_encode_cv — change Series
  dtype from np.float32 to float to avoid incompatible iloc dtype assignment
- src/api/main.py: replace deprecated @app.on_event("startup") with lifespan
  context manager; load feature names from model.feature_name_ (not JSON
  which gets overwritten by test runs); cast request field values to float
  to prevent string DeviceType overriding numeric default in _build_row
- predict.py: fill missing numeric_cols as NaN before preprocessor.transform
  so partial inference dicts don't fail the sklearn imputer
- tests/test_api.py: use TestClient as context manager (module fixture) to
  trigger lifespan startup event; fix class tests to accept client fixture

+++ src/features/build_features.py
@@ -257,7 +257,7 @@ class FeatureEngineer:
 
     def _target_encode_cv(self, df: pd.DataFrame, col: str, target: str = "isFraud") -> pd.Series:
         """CV target encoding to prevent data leakage (notebook section 9)."""
-        encoded = pd.Series(np.nan, index=df.index, dtype=np.float32)
+        encoded = pd.Series(np.nan, index=df.index, dtype=float)
         global_mean = float(df[target].mean())
         self._global_means[col] = global_mean
         kf = KFold(n_splits=self.n_splits, shuffle=False)
